+++ internal/resource/rlimit_as_enforce.go
@@ -1,0 +1,9 @@
+//go:build unix && !darwin && !openbsd
+
+package resource
+
+import "golang.org/x/sys/unix"
+
+func withSoftAS(soft uint64, fn func() error) error {
+	return withSoftRlimit(unix.RLIMIT_AS, soft, "AS", fn)
+}