+++ internal/resource/rlimit_as_res_openbsd_test.go
@@ -1,0 +1,7 @@
+//go:build openbsd
+
+package resource_test
+
+import "golang.org/x/sys/unix"
+
+func asLimitResource() int { return unix.RLIMIT_DATA }